De-bork the code base (dos2unix messed up libs and binaries - bad stuff, man.)
git-svn-id: https://openitg.svn.sourceforge.net/svnroot/openitg@607 83fadc84-e282-4d84-a09a-c4228d6ae7e5

+++ src/archutils/Unix/X11Helper.cpp
@@ -146,7 +146,7 @@ int protoErrorCallback( Display *d, XErrorEvent *err )
 	char errText[512];
 	XGetErrorText( d,  err->error_code, errText, 512 );
 	LOG->Warn( "X11 Protocol error %s (%d) has occurred, caused by request %d,%d, resource ID %d",
-		errText, err->error_code, err->request_code, err->minor_code, err->resourceid );
+		errText, (int)err->error_code, (int)err->request_code, (int)err->minor_code, (int)err->resourceid );
 
 	return 0; // Xlib ignores our return value
 }
